+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 FEATURES:
 * Add smallstep_device resource and data source.
-* Add smallstep_account resource.
+* Add smallstep_account resource and data source.
 
 CHANGES:
 * Use Smallstep API version v2025-01-01.